Retain original Gray LUT stages for deferred image transfer
Admit proved Gray XYZ/Lab LUT8, LUT16 and v4 mAB sources through
the existing unquantized image path, including palette/tint alternates.
Keep the original Gray tone curve available when the rendering intent
does not select a declared LUT. Masked and multichannel LUT images
retain their existing deferred-transfer refusal boundaries.

Update the CMS pin for exact parametric coefficients, normalized curve
clipping and original Gray LUT stages. Adapt the explicit parametric
test vector to the CMS's binary64 representation. Full-domain controls
cover both PCS families, intent selection and missing-curve refusals.

Validation: formatting, strict Clippy and 621 interpreter/parser library
tests; combined engine 973 tests, 573 release cases and Reader common
with three repetitions. Normal CMS resolution retains the same package
versions, features and dependency edges.

+++ hayro-interpret/src/color/icc/channel_tests.rs
@@ -13,7 +13,7 @@ fn unequal_channels_match_independent_gamma_controls_across_intents_and_consumer
                 Some(if version == 2 {
                     ToneReprCurve::Lut(vec![gamma * 256])
                 } else {
-                    ToneReprCurve::Parametric(vec![f32::from(gamma)])
+                    ToneReprCurve::Parametric(vec![f64::from(gamma)])
                 })
             });
             let mut bytes = source.encode().unwrap();
